doctor_app/requirements.txt
Flask
pandas
matplotlib
scikit-learn
numpy
seaborn
glob2
opencv-contrib-python-headless
Pillow
pydicom
scikit-image
Werkzeug
Flask-SQLAlchemy
tensorflow-hub
protobuf==3.20.*
python-dotenv
requests
azure-cognitiveservices-vision-customvision
azure-common
azure-nspkg==3.0.2
azure-storage==0.36.0
azure-storage-blob==12.5.0
certifi
charset-normalizer
click
colorama
Jinja2
msrest
requests-oauthlib
